+++ kr8s/objects.py
@@ -1,4 +1,4 @@
-from ._asyncio import run_sync, sync
+from ._io import run_sync, sync
 from ._objects import (
     APIObject as _APIObject,
 )
