--- kr8s/tests/test_asyncio.py
@@ -1,10 +1,0 @@
-import anyio
-
-from kr8s._asyncio import NamedTemporaryFile
-
-
-async def test_tempfiles():
-    async with NamedTemporaryFile() as f:
-        assert await f.exists()
-        assert isinstance(f, anyio.Path)
-    assert not await f.exists()
